Created login functions, directory fix, error handlers and bug fixes
Done a bunch of bug fixes making sure that files were called correctly and in the right directories. Moved a bunch of the function files to a new folder instead of keeping them in the database folder, fixed some syntax.

MAJOR: Added the login function files, called the functions on the login.php site and enabled connection to database for login. Testing is needing to be done on my end, and there is probably a few bugs.