save_election.php
This file takes the election form data created from ‘admin_create_election.php’ and saves the election and candidates into the database to create a live and existing election. The newly created and saved election can then be viewed in the dashboard. It also uploads candidate images to the uploads/ folder. It links all entries to the admin with ID 125.