Add files via upload
election_functions.php file is a helper file that stores important code for creating and showing elections and candidates. Instead of repeating the same code in many files, we put it in one place. Other admin pages can use it by including the file, making everything simpler and more organised.