Fixed Blind Indexing Bugs and Schema Bugs
Issue was within mysqli binding where I wasn't binding into the correct data convention. Needed to be as a string as its default. I also fixed the schema for what is apparently "efficiency" though I am not sure why but apparently VARBINARY works better with this data compared to BLOB. And also, I changed the indexing within the schema for guaranteed uniqueness and is in itself an index