AdminLogin Functionality Complete
Complete login functionality, as well as a new landing page for staff logon. It leads to the admin dashboard with a successfully created session

Had a few bug fixes as well, like fixing the contact page > login redirect.

A suggestion for future security is as I say admin session timeouts and regeneration logic