Rate Limiting and Admin Logging functionality
As said in the title, I added rate limiting based on IP, and even the details of the action taken. Added audit logs for the administrator side and changes admins have made and the function for it. All it needs is to be called during an action function for the admin, so keep this in mind. Refer to my comment in the security.sn.php code for the calling line.