Tally Votes #2
Ensured MVC conventions are followed as we did before.
Added the getElectionResults function to handle fetching of data.
Updated session data and made a controller so logic isnt on the front end

Hopefully all of this works, but yeah... Either way, more security is needed. And a suggestion is to disable tallying when a tally has already been completed, or a election is closes. Automatic tallying on completion time of election.