Voting back-end functionality with MVC
You get the gist of it, this includes all voting functionality (that should work). Hasn't gone through testing as user_election.php doesn't hold the form that sends the info to the controller. This also does the user logging for rate limiting and such. Will work out the kinks soon.