Massive Error Fix on Election
Added duplicate checking functions, updated addcandidate and other related functions to improve checking capabilities to differentiate between updating and CREATING. Updating candidates and election would essentially delete the existing candidates which wouldnt work if they already collected votes. Changed it so it can differentiate, and when there is a vote or you are updating candidates, they wont duplicate and names can be updated, and only deleting candidates with votes no longer work, which is good.

Added duplicate check in front-end and back-end and schema.

Known issues: Can duplicate vote in front-end, and when there are no votes on a candidate, no static paragraph is given inside the front-end. Just an error.